# services/user-management/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Docker DP - User Management",
    "description": "Docker DP - User Management API"
  },
  "host": "localhost:8080",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/dev/test-connection": {
      "post": {
        "summary": "Tests the database connection",
        "operationId": "testDatabaseConnection",
        "tags": [
          "Dev"
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "schema": {
              "$ref": "#definitions/SuccessResponse"
            }
          },
          "500": {
            "description": "An unsuccessful response",
            "schema": {
              "$ref": "#definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "Get all users",
        "operationId": "getAllUsers",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "An array of users",
            "schema": {
              "$ref": "#definitions/Users"
            }
          },
          "500": {
            "description": "Unexpected error response",
            "schema": {
              "$ref": "#definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "summary": "Create a user",
        "operationId": "createUser",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "schema": {
              "$ref": "#definitions/SaveUserCommand"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The created user",
            "schema": {
              "$ref": "#definitions/User"
            }
          },
          "500": {
            "description": "Unexpected error response",
            "schema": {
              "$ref": "#definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/api/user/{userId}": {
      "get": {
        "summary": "Get a user by id",
        "operationId": "getUserById",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "The id of the user to retrieve",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The retrieved user",
            "schema": {
              "$ref": "#definitions/User"
            }
          },
          "500": {
            "description": "Unexpected error response",
            "schema": {
              "$ref": "#definitions/ErrorResponse"
            }
          }
        }
      },
      "patch": {
        "summary": "Update a user",
        "operationId": "updateUser",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "The id of the user to update",
            "type": "string"
          },
          {
            "name": "request",
            "in": "body",
            "schema": {
              "$ref": "#definitions/UpdateUserCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response when user has been updated"
          },
          "500": {
            "description": "Unexpected error response",
            "schema": {
              "$ref": "#definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "operationId": "deleteUser",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "The id of the user to delete",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response when user has been deleted"
          },
          "500": {
            "description": "Unexpected error response",
            "schema": {
              "$ref": "#definitions/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "required": [
        "id",
        "name",
        "surname",
        "email",
        "phoneNumber"
      ],
      "properties": {
        "id": {
          "type": "number",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        }
      }
    },
    "Users": {
      "type": "array",
      "items": {
        "$ref": "#definitions/User"
      }
    },
    "SaveUserCommand": {
      "type": "object",
      "required": [
        "name",
        "surname",
        "email",
        "phoneNumber"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        }
      }
    },
    "UpdateUserCommand": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        }
      }
    },
    "ErrorResponse": {
      "$ref": "#definitions/MessageResponse"
    },
    "SuccessResponse": {
      "$ref": "#definitions/MessageResponse"
    },
    "MessageResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  }
}